+++ src/main/java/com/xmu/edu/concert_ticketseckill_system/controller/vo/OrderVo.java
@@ -10,9 +10,9 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 @Data
 public class OrderVo {
-    private int orderId;
-    private int userId;
-    private int concertId;
+    private long orderId;
+    private long userId;
+    private long concertId;
     private int ticketNum;
     private BigDecimal orderAmount;
 }
